Update intro modal
